requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
boto3>=1.34.0
python-multipart>=0.0.6
